LayoutTests/media/media-source/media-source-effectiveplaybackrate-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
Append a media segment.
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(0)))
EVENT(update)
RUN(video.play())
EXPECTED (internals.elementEffectivePlaybackRate(video) != '0') OK
END OF TEST
